Extract FeedbackRepositoryImpl
